moved utils and  added watchPlugins and edited tsconfig include
