+++ jest.config.js
@@ -17,6 +17,10 @@ const customJestConfig = {
 	},
 	preset: 'ts-jest',
 	testEnvironment: 'jest-environment-jsdom',
+	watchPlugins: [
+		'jest-watch-typeahead/filename',
+		'jest-watch-typeahead/testname',
+	],
 };
 
 // createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
