+++ package.json
@@ -95,6 +95,7 @@
 		"eslint-plugin-unused-imports": "^2.0.0",
 		"jest": "^28.1.2",
 		"jest-environment-jsdom": "^28.1.2",
+		"jest-watch-typeahead": "^2.0.0",
 		"prettier": "^2.4.1",
 		"ts-jest": "^28.0.5",
 		"typescript": "4.4.4"
