+++ src/example.test.tsx
@@ -1,4 +1,4 @@
-import { render } from 'src/test/utils';
+import { render } from '@/tests/utils';
 import Header from './components/Header/Header';
 
 jest.mock('next/router', () => ({
@@ -15,6 +15,5 @@ jest.mock('next/router', () => ({
 describe('HomePage', () => {
 	it('should render App', () => {
 		const { debug } = render(<Header />);
-		debug();
 	});
 });
